exmoxtaskmanager/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
gunicorn==20.0.4
whitenoise==5.1.0
psycopg2-binary==2.8.3